+++ example/lib/main.dart
@@ -3,6 +3,10 @@ import 'package:flutter/material.dart';
 import './onboarding_example.dart';
 
 void main() {
-  runApp(OnboardingExample());
-//  runApp(RouteExample());
+  runApp(
+    const MaterialApp(
+      debugShowCheckedModeBanner: false,
+      home: OnboardingExample(),
+    ),
+  );
 }