+++ example/lib/onboarding_example.dart
@@ -1,6 +1,59 @@
 import 'package:concentric_transition/concentric_transition.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+
+final pages = [
+  const PageData(
+    icon: Icons.format_size,
+    title: "Choose your\ninterests",
+    textColor: Colors.white,
+    bgColor: Color(0xFFFDBFDD),
+  ),
+  const PageData(
+    icon: Icons.hdr_weak,
+    title: "Drag and\ndrop to move",
+    bgColor: Color(0xFFFFFFFF),
+  ),
+  const PageData(
+    icon: Icons.bubble_chart,
+    title: "Local news\nstories",
+    bgColor: Color(0xFF0043D0),
+    textColor: Colors.white,
+  ),
+];
+
+class OnboardingExample extends StatelessWidget {
+  const OnboardingExample({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final screenWidth = MediaQuery.of(context).size.width;
+    return Scaffold(
+      body: ConcentricPageView(
+        colors: pages.map((p) => p.bgColor).toList(),
+        radius: screenWidth * 0.12,
+        curve: Curves.ease,
+        nextButtonBuilder: (context) => Icon(
+          Icons.navigate_next,
+          size: screenWidth * 0.1,
+        ),
+        // itemCount: pages.length,
+        duration: const Duration(milliseconds: 1500),
+        // opacityFactor: 2.0,
+        // scaleFactor: 0.2,
+        // verticalPosition: 0.7,
+        // direction: Axis.vertical,
+        // itemCount: pages.length,
+        // physics: NeverScrollableScrollPhysics(),
+        itemBuilder: (index) {
+          final page = pages[index % pages.length];
+          return SafeArea(
+            child: _Page(page: page),
+          );
+        },
+      ),
+    );
+  }
+}
 
 class PageData {
   final String? title;
@@ -8,7 +61,7 @@ class PageData {
   final Color bgColor;
   final Color textColor;
 
-  PageData({
+  const PageData({
     this.title,
     this.icon,
     this.bgColor = Colors.white,
@@ -16,144 +69,101 @@ class PageData {
   });
 }
 
-class OnboardingExample extends StatelessWidget {
-  final List<PageData> pages = [
-    PageData(
-      icon: Icons.format_size,
-      title: "Choose your\ninterests",
-      textColor: Colors.white,
-      bgColor: Color(0xFFFDBFDD),
-    ),
-    PageData(
-      icon: Icons.hdr_weak,
-      title: "Drag and\ndrop to move",
-      bgColor: Color(0xFFFFFFFF),
-    ),
-    PageData(
-      icon: Icons.bubble_chart,
-      title: "Local news\nstories",
-      bgColor: Color(0xFF0043D0),
-      textColor: Colors.white,
-    ),
-  ];
+class _Page extends StatelessWidget {
+  final PageData page;
 
-  List<Color> get colors => pages.map((p) => p.bgColor).toList();
+  const _Page({Key? key, required this.page}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      home: Scaffold(
-        body: ConcentricPageView(
-          colors: colors,
-//          opacityFactor: 1.0,
-//          scaleFactor: 0.0,
-          radius: 30,
-          curve: Curves.ease,
-          nextButtonContent: Icon(Icons.navigate_next),
-          finishButtonContent: Icon(Icons.done),
-          itemCount: pages.length,
-          duration: Duration(seconds: 2),
-//          verticalPosition: 0.7,
-//          direction: Axis.vertical,
-//          itemCount: pages.length,
-//          physics: NeverScrollableScrollPhysics(),
-          itemBuilder: (index, value) {
-            PageData page = pages[index % pages.length];
-            // For example scale or transform some widget by [value] param
-            //            double scale = (1 - (value.abs() * 0.4)).clamp(0.0, 1.0);
-            return Container(
-              child: Theme(
-                data: ThemeData(
-                  textTheme: TextTheme(
-                    headline6: TextStyle(
-                      color: page.textColor,
-                      fontWeight: FontWeight.w600,
-                      fontFamily: 'Helvetica',
-                      letterSpacing: 0.0,
-                      fontSize: 17,
-                    ),
-                    subtitle2: TextStyle(
-                      color: page.textColor,
-                      fontWeight: FontWeight.w300,
-                      fontSize: 18,
-                    ),
-                  ),
-                ),
-                child: PageCard(page: page),
-              ),
-            );
-          },
+    final screenHeight = MediaQuery.of(context).size.height;
+    space(double p) => SizedBox(height: screenHeight * p / 100);
+    return Column(
+      children: [
+        space(10),
+        _Image(page: page),
+        space(8),
+        _Text(
+          page: page,
+          style: TextStyle(
+            fontSize: screenHeight * 0.046,
+          ),
         ),
-      ),
+      ],
     );
   }
 }
 
-class PageCard extends StatelessWidget {
-  final PageData page;
-
-  const PageCard({
+class _Text extends StatelessWidget {
+  const _Text({
     Key? key,
     required this.page,
+    this.style,
   }) : super(key: key);
+
+  final PageData page;
+  final TextStyle? style;
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      margin: EdgeInsets.symmetric(
-        horizontal: 30.0,
-      ),
-      child: Column(
-//        mainAxisAlignment: MainAxisAlignment.center,
-        children: <Widget>[
-          _buildPicture(context),
-          SizedBox(height: 30),
-          _buildText(context),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildText(BuildContext context) {
     return Text(
-      page.title!,
-      style: Theme.of(context).textTheme.headline6,
+      page.title ?? '',
+      style: TextStyle(
+        color: page.textColor,
+        fontWeight: FontWeight.w600,
+        fontFamily: 'Helvetica',
+        letterSpacing: 0.0,
+        fontSize: 18,
+        height: 1.2,
+      ).merge(style),
       textAlign: TextAlign.center,
     );
   }
+}
 
-  Widget _buildPicture(
-    BuildContext context, {
-    double size = 190,
-    double iconSize = 170,
-  }) {
+class _Image extends StatelessWidget {
+  const _Image({
+    Key? key,
+    required this.page,
+    this.size = 190,
+    this.iconSize = 170,
+  }) : super(key: key);
+
+  final PageData page;
+  final double size;
+  final double iconSize;
+
+  @override
+  Widget build(BuildContext context) {
+    final bgColor = page.bgColor
+        // .withBlue(page.bgColor.blue - 40)
+        .withGreen(page.bgColor.green + 20)
+        .withRed(page.bgColor.red - 100)
+        .withAlpha(90);
+
+    final icon1Color =
+        page.bgColor.withBlue(page.bgColor.blue - 10).withGreen(220);
+    final icon2Color = page.bgColor.withGreen(66).withRed(77);
+    final icon3Color = page.bgColor.withRed(111).withGreen(220);
+
     return Container(
       width: size,
       height: size,
       decoration: BoxDecoration(
-        borderRadius: BorderRadius.all(Radius.circular(60.0)),
-        color: page.bgColor
-//            .withBlue(page.bgColor.blue - 40)
-            .withGreen(page.bgColor.green + 20)
-            .withRed(page.bgColor.red - 100)
-            .withAlpha(90),
-      ),
-      margin: EdgeInsets.only(
-        top: 140,
+        borderRadius: const BorderRadius.all(Radius.circular(60.0)),
+        color: bgColor,
       ),
       child: Stack(
+        clipBehavior: Clip.none,
         fit: StackFit.expand,
-        children: <Widget>[
+        children: [
           Positioned.fill(
             child: RotatedBox(
               quarterTurns: 2,
               child: Icon(
                 page.icon,
                 size: iconSize + 20,
-                color: page.bgColor
-                    .withBlue(page.bgColor.blue - 10)
-                    .withGreen(220),
+                color: icon1Color,
               ),
             ),
             right: -5,
@@ -165,14 +175,14 @@ class PageCard extends StatelessWidget {
               child: Icon(
                 page.icon,
                 size: iconSize + 20,
-                color: page.bgColor.withGreen(66).withRed(77),
+                color: icon2Color,
               ),
             ),
           ),
           Icon(
             page.icon,
             size: iconSize,
-            color: page.bgColor.withRed(111).withGreen(220),
+            color: icon3Color,
           ),
         ],
       ),