+++ example/lib/route_example.dart
@@ -2,6 +2,8 @@ import 'package:concentric_transition/concentric_transition.dart';
 import 'package:flutter/material.dart';
 
 class RouteExample extends StatefulWidget {
+  const RouteExample({Key? key}) : super(key: key);
+
   @override
   _RouteExampleState createState() => _RouteExampleState();
 }
@@ -9,13 +11,15 @@ class RouteExample extends StatefulWidget {
 class _RouteExampleState extends State<RouteExample> {
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return const MaterialApp(
       home: Page1(),
     );
   }
 }
 
 class Page1 extends StatelessWidget {
+  const Page1({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -23,10 +27,10 @@ class Page1 extends StatelessWidget {
 //      appBar: AppBar(title: Text("Page 1")),
       body: Center(
         child: MaterialButton(
-          child: Text("Next"),
+          child: const Text("Next"),
           onPressed: () {
             Navigator.push(context, ConcentricPageRoute(builder: (ctx) {
-              return Page2();
+              return const Page2();
             }));
           },
         ),
@@ -36,6 +40,8 @@ class Page1 extends StatelessWidget {
 }
 
 class Page2 extends StatelessWidget {
+  const Page2({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -43,7 +49,7 @@ class Page2 extends StatelessWidget {
 //      appBar: AppBar(title: Text("Page 2")),
       body: Center(
         child: MaterialButton(
-          child: Text("Back"),
+          child: const Text("Back"),
           onPressed: () => Navigator.pop(context),
         ),
       ),