+++ lib/page_route.dart
@@ -1,5 +1,5 @@
 import 'package:concentric_transition/clipper.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 
 class ConcentricPageRoute<T> extends PageRoute<T> {
   ConcentricPageRoute({
@@ -84,13 +84,13 @@ class _FadeInPageTransition extends StatelessWidget {
   })  : _opacityAnimation = routeAnimation.drive(_easeInTween),
         super(key: key);
 
-//  // Fractional offset from 1/4 screen below the top to fully on screen.
-//  static final Tween<Offset> _bottomUpTween = Tween<Offset>(
-//    begin: const Offset(0.0, 0.25),
-//    end: Offset.zero,
-//  );
-//  static final Animatable<double> _fastOutSlowInTween =
-//      CurveTween(curve: Curves.fastOutSlowIn);
+  // // Fractional offset from 1/4 screen below the top to fully on screen.
+  // static final Tween<Offset> _bottomUpTween = Tween<Offset>(
+  //   begin: const Offset(0.0, 0.25),
+  //   end: Offset.zero,
+  // );
+  // static final Animatable<double> _fastOutSlowInTween =
+  //     CurveTween(curve: Curves.fastOutSlowIn);
   static final Animatable<double> _easeInTween =
       CurveTween(curve: Curves.easeIn);
 
@@ -99,15 +99,13 @@ class _FadeInPageTransition extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-//    return FadeTransition(
-//      opacity: _opacityAnimation,
-//      child: child,
-//    );
-    return Container(
-      child: ClipPath(
-        clipper: ConcentricClipper(progress: _opacityAnimation.value),
-        child: child,
-      ),
+    // return FadeTransition(
+    //   opacity: _opacityAnimation,
+    //   child: child,
+    // );
+    return ClipPath(
+      clipper: ConcentricClipper(progress: _opacityAnimation.value),
+      child: child,
     );
   }
 }