+++ lib/page_view.dart
@@ -1,13 +1,12 @@
-import 'dart:async';
-
-import 'package:flutter/material.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/rendering.dart';
+import 'package:flutter/widgets.dart';
 
 import 'clipper.dart';
 
 class ConcentricPageView extends StatefulWidget {
   /// The [value] will help to provide some animations
-  final Function(int index, double value) itemBuilder;
+  final Function(int index) itemBuilder;
   final Function(int page)? onChange;
   final Function? onFinish;
   final int? itemCount;
@@ -26,10 +25,7 @@ class ConcentricPageView extends StatefulWidget {
   final Curve curve;
 
   /// Useful for adding a next icon to the page view button
-  final Widget? nextButtonContent;
-
-  /// Useful for adding a complete icon to the page view button on final page
-  final Widget? finishButtonContent;
+  final WidgetBuilder? nextButtonBuilder;
 
   const ConcentricPageView({
     Key? key,
@@ -47,13 +43,10 @@ class ConcentricPageView extends StatefulWidget {
     this.radius = 40.0,
     this.verticalPosition = 0.75,
     this.direction = Axis.horizontal,
-//    this.physics = const NeverScrollableScrollPhysics(),
-    this.physics,
+    this.physics = const ClampingScrollPhysics(),
     this.duration = const Duration(milliseconds: 1500),
-    this.curve = Curves.easeOutSine,
-    this.finishButtonContent,
-    // Cubic(0.7, 0.5, 0.5, 0.1)
-    this.nextButtonContent,
+    this.curve = Curves.easeOutSine, // const Cubic(0.7, 0.5, 0.5, 0.1),
+    this.nextButtonBuilder,
   })  : assert(colors.length >= 2),
         super(key: key);
 
@@ -62,8 +55,7 @@ class ConcentricPageView extends StatefulWidget {
 }
 
 class _ConcentricPageViewState extends State<ConcentricPageView> {
-  PageController? _pageController;
-
+  late PageController _pageController;
   double _progress = 0;
   int _prevPage = 0;
   Color? _prevColor;
@@ -73,171 +65,124 @@ class _ConcentricPageViewState extends State<ConcentricPageView> {
   void initState() {
     _prevColor = widget.colors[_prevPage];
     _nextColor = widget.colors[_prevPage + 1];
-    _pageController = widget.pageController != null
-        ? widget.pageController
-        : PageController(
-            initialPage: 0,
-          );
-
-    _pageController!.addListener(_onScroll);
+    _pageController = (widget.pageController ?? PageController(initialPage: 0))
+      ..addListener(_onScroll);
     super.initState();
   }
 
   @override
   void dispose() {
-    _pageController!.removeListener(_onScroll);
-    if (widget.pageController == null) {
-      _pageController!.dispose();
-    }
+    _pageController.removeListener(_onScroll);
+    _pageController.dispose();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     return Stack(
+      clipBehavior: Clip.none,
       alignment: Alignment.center,
-      children: <Widget>[
-        AnimatedBuilder(
-          animation: _pageController!,
-          builder: (ctx, _) {
-            return Container(
-              color: _prevColor, // Colors.white,
-              child: ClipPath(
-                clipper: ConcentricClipper(
-                  progress: _progress,
-                  reverse: widget.reverse,
-                  radius: widget.radius,
-                  verticalPosition: widget.verticalPosition,
-                ),
-                child: Container(
-                  color: _nextColor,
-//                  color: ColorTween(begin: _prevColor, end: _nextColor)
-//                      .transform(_progress), // Colors.blue,
-                ),
-              ),
-            );
-          },
-        ),
-        PageView.builder(
-          //          onPageChanged: (page) {
-          //            print('new page $page');
-          //          },
-          controller: _pageController,
-          reverse: widget.reverse,
-          physics: widget.physics,
-          scrollDirection: widget.direction,
-          itemCount: widget.itemCount,
-          pageSnapping: widget.pageSnapping,
-          onPageChanged: _onPageChanged,
-          itemBuilder: (context, index) {
-//            var i = index % widget.children.length;
-            return AnimatedBuilder(
-              animation: _pageController!,
-              builder: (BuildContext context, child) {
-                // on the first render, the pageController.page is null,
-                // this is a dirty hack
-                if (!_pageController!.position.hasContentDimensions) {
-                  Future.delayed(Duration(microseconds: 1), () {
-                    setState(() {});
-                  });
-                  return Container();
-                }
-
-                final double value = _pageController!.page! - index;
-                final double scale =
-                    (1 - (value.abs() * widget.scaleFactor)).clamp(0.0, 1.0);
-                final double opacity =
-                    (1 - (value.abs() * widget.opacityFactor)).clamp(0.0, 1.0);
-
-                return Transform.scale(
-                  scale: scale,
-                  child: Opacity(
-//                    duration: Duration(milliseconds: 1000),
-                    opacity: opacity,
-                    child: widget.itemBuilder(index, value),
-                  ),
-                );
-              },
-            );
-          },
-        ),
-        StatefulBuilder(
-          builder: (BuildContext context, StateSetter setState) {
-            _pageController!.addListener(() {
-              if (_pageController!.page is int) {
-                setState(() {});
-              }
-            });
-
-            return Positioned(
-                top: MediaQuery.of(context).size.height *
-                    widget.verticalPosition,
-                child: TweenAnimationBuilder(
-                  key: Key('${_pageController!.page?.floor()}'),
-                  tween: Tween<double>(begin: 0.0, end: 1.0),
-                  duration: Duration(
-                      milliseconds:
-                          (widget.duration.inMilliseconds / 2).round()),
-                  builder: (BuildContext context, double value, Widget? child) {
-                    return Opacity(
-                      opacity: value,
-                      child: child,
-                    );
-                  },
-                  child: _buildButton(),
-                ));
-          },
+      children: [
+        _buildClipper(),
+        _buildPage(),
+        Positioned(
+          top: MediaQuery.of(context).size.height * widget.verticalPosition,
+          child: _Button(
+            pageController: _pageController,
+            widget: widget,
+          ),
         ),
       ],
     );
   }
 
-  Widget _buildButton() {
-    var isFinal = _pageController!.page == widget.colors.length - 1;
-
-    return RawMaterialButton(
-      onPressed: () {
-        if (isFinal) {
-          if (widget.onFinish != null) {
-            widget.onFinish!();
-          }
-        } else {
-          _pageController!.nextPage(
-            duration: widget.duration,
-            curve: widget.curve,
-          );
+  Widget _buildPage() {
+    return PageView.builder(
+      scrollBehavior: ScrollConfiguration.of(context).copyWith(
+        scrollbars: false,
+        overscroll: false,
+        dragDevices: {
+          PointerDeviceKind.touch,
+          PointerDeviceKind.mouse,
+        },
+      ),
+      clipBehavior: Clip.none,
+      scrollDirection: widget.direction,
+      controller: _pageController,
+      reverse: widget.reverse,
+      physics: widget.physics,
+      itemCount: widget.itemCount,
+      pageSnapping: widget.pageSnapping,
+      onPageChanged: (int page) {
+        if (widget.onChange != null) {
+          widget.onChange!(page);
         }
       },
-      constraints: BoxConstraints(
-        minWidth: widget.radius * 2,
-        minHeight: widget.radius * 2,
-      ),
-      shape: CircleBorder(),
-      child: isFinal ? widget.finishButtonContent : widget.nextButtonContent,
+      itemBuilder: (context, index) {
+        if (!_pageController.position.hasContentDimensions) {
+          return const SizedBox();
+        }
+        return AnimatedBuilder(
+          animation: _pageController,
+          child: widget.itemBuilder(index),
+          builder: (context, child) {
+            final progress = _pageController.page! - index;
+            if (widget.opacityFactor != 0) {
+              child = Opacity(
+                opacity: (1 - (progress.abs() * widget.opacityFactor))
+                    .clamp(0.0, 1.0),
+                child: child,
+              );
+            }
+            if (widget.scaleFactor != 0) {
+              child = Transform.scale(
+                scale:
+                    (1 - (progress.abs() * widget.scaleFactor)).clamp(0.0, 1.0),
+                child: child,
+              );
+            }
+            return child!;
+          },
+        );
+      },
     );
   }
 
-  void _onPageChanged(int page) {
-    if (widget.onChange != null) {
-      widget.onChange!(page);
-    }
+  Widget _buildClipper() {
+    return AnimatedBuilder(
+      animation: _pageController,
+      builder: (ctx, _) {
+        return ColoredBox(
+          color: _prevColor!,
+          child: ClipPath(
+            clipper: ConcentricClipper(
+              progress: _progress,
+              reverse: widget.reverse,
+              radius: widget.radius,
+              verticalPosition: widget.verticalPosition,
+            ),
+            child: ColoredBox(
+              color: _nextColor!,
+              child: const SizedBox.expand(),
+            ),
+          ),
+        );
+      },
+    );
   }
 
   void _onScroll() {
-    ScrollDirection direction = _pageController!.position.userScrollDirection;
+    final direction = _pageController.position.userScrollDirection;
     if (direction == ScrollDirection.forward) {
-      _prevPage = (_pageController!.page! + 0.001).toInt();
-      _progress = _pageController!.page! - _prevPage;
-//    } else if (direction == ScrollDirection.reverse) {
+      _prevPage = (_pageController.page! + 0.0001).toInt();
+      _progress = _pageController.page! - _prevPage;
     } else {
-      _prevPage = (_pageController!.page! - 0.001).toInt();
-      _progress = _pageController!.page! - _prevPage;
-//    } else {
-//      _progress = 0;
+      _prevPage = (_pageController.page! - 0.0001).toInt();
+      _progress = _pageController.page! - _prevPage;
     }
 
-    final int total = widget.colors.length;
-    int prevIndex = _prevPage % total;
+    final total = widget.colors.length;
+    final prevIndex = _prevPage % total;
     int nextIndex = prevIndex + 1;
 
     if (prevIndex == total - 1) {
@@ -247,6 +192,62 @@ class _ConcentricPageViewState extends State<ConcentricPageView> {
     _prevColor = widget.colors[prevIndex];
     _nextColor = widget.colors[nextIndex];
 
-    widget.notifier?.value = _pageController!.page! - _prevPage;
+    widget.notifier?.value = _pageController.page! - _prevPage;
+  }
+}
+
+class _Button extends StatelessWidget {
+  const _Button({
+    Key? key,
+    required this.pageController,
+    required this.widget,
+  }) : super(key: key);
+
+  final PageController pageController;
+  final ConcentricPageView widget;
+
+  @override
+  Widget build(BuildContext context) {
+    final size = widget.radius * 2;
+    Widget? child = widget.nextButtonBuilder != null
+        ? widget.nextButtonBuilder!(context)
+        : null;
+
+    child = GestureDetector(
+      excludeFromSemantics: true,
+      onTap: () {
+        final isFinal = pageController.page == widget.colors.length - 1;
+        if (isFinal && widget.onFinish != null) {
+          widget.onFinish!();
+          return;
+        }
+        pageController.nextPage(
+          duration: widget.duration,
+          curve: widget.curve,
+        );
+      },
+      child: DecoratedBox(
+        decoration: const BoxDecoration(shape: BoxShape.circle),
+        child: SizedBox.fromSize(
+          child: child,
+          size: Size.square(size),
+        ),
+      ),
+    );
+
+    return AnimatedBuilder(
+      animation: pageController,
+      builder: (context, child) {
+        final currentPage = pageController.page?.floor() ?? 0;
+        final progress = (pageController.page ?? 0) - currentPage;
+        return AnimatedOpacity(
+          opacity: progress > 0.02 ? 0.0 : 1.0,
+          curve: Curves.ease,
+          duration: const Duration(milliseconds: 150),
+          child: child!,
+        );
+      },
+      child: child,
+    );
   }
 }